pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "catchai-copiloto"
version = "1.0.0"
description = "Copiloto conversacional sobre documentos PDF usando IA"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Equipo CatchAI", email = "equipo@catchai.com"}
]
keywords = ["ia", "pdf", "conversacional", "analisis-documentos", "langchain", "gemini"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "streamlit>=1.37.0",
    "langchain>=0.2.6",
    "langchain-community>=0.2.6",
    "chromadb>=0.5.0",
    "sentence-transformers>=2.7.0",
    "pypdf>=4.3.0",
    "pdfplumber>=0.11.0",
    "python-dotenv>=1.0.0",
    "google-genai>=0.1.0",
    "tiktoken>=0.7.0",
    "numpy>=1.24.0",
    "pydantic>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/tuusuario/catchai-copiloto"
Repository = "https://github.com/tuusuario/catchai-copiloto"
Issues = "https://github.com/tuusuario/catchai-copiloto/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
